Improve email recipients and overview assets
