+++ src/Admin/Menu.php
@@ -40,7 +40,7 @@ class Menu
         );
 
         add_submenu_page('fp-dms-dashboard', __('Dashboard', 'fp-dms'), __('Dashboard', 'fp-dms'), 'manage_options', 'fp-dms-dashboard', [DashboardPage::class, 'render']);
-        add_submenu_page('fp-dms-dashboard', __('Overview', 'fp-dms'), __('Overview', 'fp-dms'), 'manage_options', 'fp-dms-overview', [OverviewPage::class, 'render']);
+        $overviewHook = add_submenu_page('fp-dms-dashboard', __('Overview', 'fp-dms'), __('Overview', 'fp-dms'), 'manage_options', 'fp-dms-overview', [OverviewPage::class, 'render']);
         add_submenu_page('fp-dms-dashboard', __('Clients', 'fp-dms'), __('Clients', 'fp-dms'), 'manage_options', 'fp-dms-clients', [ClientsPage::class, 'render']);
         add_submenu_page('fp-dms-dashboard', __('Data Sources', 'fp-dms'), __('Data Sources', 'fp-dms'), 'manage_options', 'fp-dms-datasources', [DataSourcesPage::class, 'render']);
         add_submenu_page('fp-dms-dashboard', __('Schedules', 'fp-dms'), __('Schedules', 'fp-dms'), 'manage_options', 'fp-dms-schedules', [SchedulesPage::class, 'render']);
@@ -53,6 +53,10 @@ class Menu
 
         if ($hook) {
             add_action('load-' . $hook, [self::class, 'enqueue_assets']);
+        }
+
+        if (! empty($overviewHook)) {
+            OverviewPage::registerAssetsHook($overviewHook);
         }
     }
